+++ libs/util/lia-exporter/src/index.ts
@@ -1,2 +1,1 @@
 export * from "./lib/lia-exporter";
-export * from "./lib/types";
